requirements.txt
flask>=3.0.0
openai>=1.50.0
gunicorn>=21.2.0
requests>=2.31.0
markdown>=3.5.1
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0

